requirements.txt
pypdf>=3.0.0
pandas>=1.5.0
openpyxl>=3.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-mistralai>=0.0.11
mistralai>=0.0.12
requests>=2.31.0
